requirements.txt
langchain>=0.1.20
cohere
chromadb
PyPDF2
streamlit
python-dotenv